Splitting out files, renaming to be upper-case.
