+++ common.h
@@ -1,6 +1,8 @@
 #ifndef __COMMON_H
 #define __COMMON_H
 
+#include "pico/stdlib.h"
+
 typedef unsigned int size_t;
 
 #endif
