+++ e-ink.cpp
@@ -1,8 +1,9 @@
-#include "led.h"
-#include "pins.h"
-#include "display.h"
-#include "font.h"
-#include "path.h"
+#include "Led.h"
+#include "Pins.h"
+#include "Display.h"
+#include "Font.h"
+#include "Path.h"
+#include "DejaVuSans.h"
 
 #include "pico/stdlib.h"
 #include "hardware/gpio.h"
@@ -24,2720 +25,6 @@
 #define RESET_PIN 7    // 10 // RESET // low for reset
 #define DC_PIN    8    // 11 // DC    // Data (high), Command (low)
 
-class DejaVuSans: public FontFace
-{
-public:
-    virtual ~DejaVuSans(void)
-    {
-    }
-
-    glyph operator[](int character) const override
-    {
-        switch (character)
-        {
-            case ' ': return space;
-            case '!': return exclam;
-            case '"': return quotedbl;
-            case '#': return numbersign;
-            case '$': return dollar;
-            case '%': return percent;
-            case '&': return ampersand;
-            case '\'': return quotesingle;
-            case '(': return parenleft;
-            case ')': return parenright;
-            case '*': return asterisk;
-            case '+': return plus;
-            case ',': return comma;
-            case '-': return hyphen;
-            case '.': return period;
-            case '/': return slash;
-            case '0': return zero;
-            case '1': return one;
-            case '2': return two;
-            case '3': return three;
-            case '4': return four;
-            case '5': return five;
-            case '6': return six;
-            case '7': return seven;
-            case '8': return eight;
-            case '9': return nine;
-            case ':': return colon;
-            case ';': return semicolon;
-            case '<': return less;
-            case '=': return equal;
-            case '>': return greater;
-            case '?': return question;
-            case '@': return at;
-            case 'A': return A;
-            case 'B': return B;
-            case 'C': return C;
-            case 'D': return D;
-            case 'E': return E;
-            case 'F': return F;
-            case 'G': return G;
-            case 'H': return H;
-            case 'I': return I;
-            case 'J': return J;
-            case 'K': return K;
-            case 'L': return L;
-            case 'M': return M;
-            case 'N': return N;
-            case 'O': return O;
-            case 'P': return P;
-            case 'Q': return Q;
-            case 'R': return R;
-            case 'S': return S;
-            case 'T': return T;
-            case 'U': return U;
-            case 'V': return V;
-            case 'W': return W;
-            case 'X': return X;
-            case 'Y': return Y;
-            case 'Z': return Z;
-            case '[': return bracketleft;
-            case '\\': return backslash;
-            case ']': return bracketright;
-            case '^': return asciicircum;
-            case '_': return underscore;
-            case '`': return grave;
-            case 'a': return a;
-            case 'b': return b;
-            case 'c': return c;
-            case 'd': return d;
-            case 'e': return e;
-            case 'f': return f;
-            case 'g': return g;
-            case 'h': return h;
-            case 'i': return i;
-            case 'j': return j;
-            case 'k': return k;
-            case 'l': return l;
-            case 'm': return m;
-            case 'n': return n;
-            case 'o': return o;
-            case 'p': return p;
-            case 'q': return q;
-            case 'r': return r;
-            case 's': return s;
-            case 't': return t;
-            case 'u': return u;
-            case 'v': return v;
-            case 'w': return w;
-            case 'x': return x;
-            case 'y': return y;
-            case 'z': return z;
-            case '{': return braceleft;
-            case '|': return bar;
-            case '}': return braceright;
-            case '~': return asciitilde;
-            default:
-                return None;
-        }
-    }
-private:
-
-    static em None(EmBox* box)
-    {
-        return 0;
-    }
-
-    static em space(EmBox* box)
-    {
-        return 651;
-    }
-
-    static em exclam(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({309.0, 254.0});
-        pathSegment0.
-            line({512.0, 254.0})->
-            line({512.0, 0.0})->
-            line({309.0, 0.0})->
-            line({309.0, 254.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({309.0, 1493.0});
-        pathSegment1.
-            line({512.0, 1493.0})->
-            line({512.0, 838.0})->
-            line({492.0, 481.0})->
-            line({330.0, 481.0})->
-            line({309.0, 838.0})->
-            line({309.0, 1493.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 821;
-    }
-
-    static em quotedbl(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({367.0, 1493.0});
-        pathSegment0.
-            line({367.0, 938.0})->
-            line({197.0, 938.0})->
-            line({197.0, 1493.0})->
-            line({367.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({745.0, 1493.0});
-        pathSegment1.
-            line({745.0, 938.0})->
-            line({575.0, 938.0})->
-            line({575.0, 1493.0})->
-            line({745.0, 1493.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 942;
-    }
-
-    static em numbersign(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1047.0, 901.0});
-        pathSegment0.
-            line({756.0, 901.0})->
-            line({672.0, 567.0})->
-            line({965.0, 567.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({897.0, 1470.0});
-        pathSegment1.
-            line({793.0, 1055.0})->
-            line({1085.0, 1055.0})->
-            line({1190.0, 1470.0})->
-            line({1350.0, 1470.0})->
-            line({1247.0, 1055.0})->
-            line({1559.0, 1055.0})->
-            line({1559.0, 901.0})->
-            line({1208.0, 901.0})->
-            line({1126.0, 567.0})->
-            line({1444.0, 567.0})->
-            line({1444.0, 414.0})->
-            line({1087.0, 414.0})->
-            line({983.0, 0.0})->
-            line({823.0, 0.0})->
-            line({926.0, 414.0})->
-            line({633.0, 414.0})->
-            line({530.0, 0.0})->
-            line({369.0, 0.0})->
-            line({473.0, 414.0})->
-            line({158.0, 414.0})->
-            line({158.0, 567.0})->
-            line({510.0, 567.0})->
-            line({594.0, 901.0})->
-            line({272.0, 901.0})->
-            line({272.0, 1055.0})->
-            line({633.0, 1055.0})->
-            line({735.0, 1470.0})->
-            line({897.0, 1470.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1716;
-    }
-
-    static em dollar(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({692.0, -301.0});
-        pathSegment0.
-            line({592.0, -301.0})->
-            line({591.0, 0.0})->
-            quadratic({486.0, 2.0}, {381.0, 24.5})->
-            quadratic({276.0, 47.0}, {170.0, 92.0})->
-            line({170.0, 272.0})->
-            quadratic({272.0, 208.0}, {376.5, 175.5})->
-            quadratic({481.0, 143.0}, {592.0, 142.0})->
-            line({592.0, 598.0})->
-            quadratic({371.0, 634.0}, {270.5, 720.0})->
-            quadratic({170.0, 806.0}, {170.0, 956.0})->
-            quadratic({170.0, 1119.0}, {279.0, 1213.0})->
-            quadratic({388.0, 1307.0}, {592.0, 1321.0})->
-            line({592.0, 1556.0})->
-            line({692.0, 1556.0})->
-            line({692.0, 1324.0})->
-            quadratic({785.0, 1320.0}, {872.0, 1304.5})->
-            quadratic({959.0, 1289.0}, {1042.0, 1262.0})->
-            line({1042.0, 1087.0})->
-            quadratic({959.0, 1129.0}, {871.5, 1152.0})->
-            quadratic({784.0, 1175.0}, {692.0, 1179.0})->
-            line({692.0, 752.0})->
-            quadratic({919.0, 717.0}, {1026.0, 627.0})->
-            quadratic({1133.0, 537.0}, {1133.0, 381.0})->
-            quadratic({1133.0, 212.0}, {1019.5, 114.5})->
-            quadratic({906.0, 17.0}, {692.0, 2.0})->
-            line({692.0, -301.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({592.0, 770.0});
-        pathSegment1.
-            line({592.0, 1180.0})->
-            quadratic({476.0, 1167.0}, {415.0, 1114.0})->
-            quadratic({354.0, 1061.0}, {354.0, 973.0})->
-            quadratic({354.0, 887.0}, {410.5, 839.0})->
-            quadratic({467.0, 791.0}, {592.0, 770.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        Path pathSegment2 = Path::start({692.0, 578.0});
-        pathSegment2.
-            line({692.0, 145.0})->
-            quadratic({819.0, 162.0}, {883.5, 217.0})->
-            quadratic({948.0, 272.0}, {948.0, 362.0})->
-            quadratic({948.0, 450.0}, {886.5, 502.0})->
-            quadratic({825.0, 554.0}, {692.0, 578.0})->
-            close(&pathSegment2);
-        pathSegment2.drawTo(box);
-
-        return 2000;
-    }
-
-    static em percent(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1489.0, 657.0});
-        pathSegment0.
-            quadratic({1402.0, 657.0}, {1352.5, 583.0})->
-            quadratic({1303.0, 509.0}, {1303.0, 377.0})->
-            quadratic({1303.0, 247.0}, {1352.5, 172.5})->
-            quadratic({1402.0, 98.0}, {1489.0, 98.0})->
-            quadratic({1574.0, 98.0}, {1623.5, 172.5})->
-            quadratic({1673.0, 247.0}, {1673.0, 377.0})->
-            quadratic({1673.0, 508.0}, {1623.5, 582.5})->
-            quadratic({1574.0, 657.0}, {1489.0, 657.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({1489.0, 784.0});
-        pathSegment1.
-            quadratic({1647.0, 784.0}, {1740.0, 674.0})->
-            quadratic({1833.0, 564.0}, {1833.0, 377.0})->
-            quadratic({1833.0, 190.0}, {1739.5, 80.5})->
-            quadratic({1646.0, -29.0}, {1489.0, -29.0})->
-            quadratic({1329.0, -29.0}, {1236.0, 80.5})->
-            quadratic({1143.0, 190.0}, {1143.0, 377.0})->
-            quadratic({1143.0, 565.0}, {1236.5, 674.5})->
-            quadratic({1330.0, 784.0}, {1489.0, 784.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        Path pathSegment2 = Path::start({457.0, 1393.0});
-        pathSegment2.
-            quadratic({371.0, 1393.0}, {321.5, 1318.5})->
-            quadratic({272.0, 1244.0}, {272.0, 1114.0})->
-            quadratic({272.0, 982.0}, {321.0, 908.0})->
-            quadratic({370.0, 834.0}, {457.0, 834.0})->
-            quadratic({544.0, 834.0}, {593.5, 908.0})->
-            quadratic({643.0, 982.0}, {643.0, 1114.0})->
-            quadratic({643.0, 1243.0}, {593.0, 1318.0})->
-            quadratic({543.0, 1393.0}, {457.0, 1393.0})->
-            close(&pathSegment2);
-        pathSegment2.drawTo(box);
-
-        Path pathSegment3 = Path::start({1360.0, 1520.0});
-        pathSegment3.
-            line({1520.0, 1520.0})->
-            line({586.0, -29.0})->
-            line({426.0, -29.0})->
-            close(&pathSegment3);
-        pathSegment3.drawTo(box);
-
-        Path pathSegment4 = Path::start({457.0, 1520.0});
-        pathSegment4.
-            quadratic({615.0, 1520.0}, {709.0, 1410.5})->
-            quadratic({803.0, 1301.0}, {803.0, 1114.0})->
-            quadratic({803.0, 925.0}, {709.5, 816.0})->
-            quadratic({616.0, 707.0}, {457.0, 707.0})->
-            quadratic({298.0, 707.0}, {205.5, 816.5})->
-            quadratic({113.0, 926.0}, {113.0, 1114.0})->
-            quadratic({113.0, 1300.0}, {206.0, 1410.0})->
-            quadratic({299.0, 1520.0}, {457.0, 1520.0})->
-            close(&pathSegment4);
-        pathSegment4.drawTo(box);
-
-        return 1946;
-    }
-
-    static em ampersand(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({498.0, 803.0});
-        pathSegment0.
-            quadratic({407.0, 722.0}, {364.5, 641.5})->
-            quadratic({322.0, 561.0}, {322.0, 473.0})->
-            quadratic({322.0, 327.0}, {428.0, 230.0})->
-            quadratic({534.0, 133.0}, {694.0, 133.0})->
-            quadratic({789.0, 133.0}, {872.0, 164.5})->
-            quadratic({955.0, 196.0}, {1028.0, 260.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({639.0, 915.0});
-        pathSegment1.
-            line({1147.0, 395.0})->
-            quadratic({1206.0, 484.0}, {1239.0, 585.5})->
-            quadratic({1272.0, 687.0}, {1278.0, 801.0})->
-            line({1464.0, 801.0})->
-            quadratic({1452.0, 669.0}, {1400.0, 540.0})->
-            quadratic({1348.0, 411.0}, {1255.0, 285.0})->
-            line({1534.0, 0.0})->
-            line({1282.0, 0.0})->
-            line({1139.0, 147.0})->
-            quadratic({1035.0, 58.0}, {921.0, 14.5})->
-            quadratic({807.0, -29.0}, {676.0, -29.0})->
-            quadratic({435.0, -29.0}, {282.0, 108.5})->
-            quadratic({129.0, 246.0}, {129.0, 461.0})->
-            quadratic({129.0, 589.0}, {196.0, 701.5})->
-            quadratic({263.0, 814.0}, {397.0, 913.0})->
-            quadratic({349.0, 976.0}, {324.0, 1038.5})->
-            quadratic({299.0, 1101.0}, {299.0, 1161.0})->
-            quadratic({299.0, 1323.0}, {410.0, 1421.5})->
-            quadratic({521.0, 1520.0}, {705.0, 1520.0})->
-            quadratic({788.0, 1520.0}, {870.5, 1502.0})->
-            quadratic({953.0, 1484.0}, {1038.0, 1448.0})->
-            line({1038.0, 1266.0})->
-            quadratic({951.0, 1313.0}, {872.0, 1337.5})->
-            quadratic({793.0, 1362.0}, {725.0, 1362.0})->
-            quadratic({620.0, 1362.0}, {554.5, 1306.5})->
-            quadratic({489.0, 1251.0}, {489.0, 1163.0})->
-            quadratic({489.0, 1112.0}, {518.5, 1060.5})->
-            quadratic({548.0, 1009.0}, {639.0, 915.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1597;
-    }
-
-    static em quotesingle(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({367.0, 1493.0});
-        pathSegment0.
-            line({367.0, 938.0})->
-            line({197.0, 938.0})->
-            line({197.0, 1493.0})->
-            line({367.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 563;
-    }
-
-    static em parenleft(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({635.0, 1554.0});
-        pathSegment0.
-            quadratic({501.0, 1324.0}, {436.0, 1099.0})->
-            quadratic({371.0, 874.0}, {371.0, 643.0})->
-            quadratic({371.0, 412.0}, {436.5, 185.5})->
-            quadratic({502.0, -41.0}, {635.0, -270.0})->
-            line({475.0, -270.0})->
-            quadratic({325.0, -35.0}, {250.5, 192.0})->
-            quadratic({176.0, 419.0}, {176.0, 643.0})->
-            quadratic({176.0, 866.0}, {250.0, 1092.0})->
-            quadratic({324.0, 1318.0}, {475.0, 1554.0})->
-            line({635.0, 1554.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 799;
-    }
-
-    static em parenright(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({164.0, 1554.0});
-        pathSegment0.
-            line({324.0, 1554.0})->
-            quadratic({474.0, 1318.0}, {548.5, 1092.0})->
-            quadratic({623.0, 866.0}, {623.0, 643.0})->
-            quadratic({623.0, 419.0}, {548.5, 192.0})->
-            quadratic({474.0, -35.0}, {324.0, -270.0})->
-            line({164.0, -270.0})->
-            quadratic({297.0, -41.0}, {362.5, 185.5})->
-            quadratic({428.0, 412.0}, {428.0, 643.0})->
-            quadratic({428.0, 874.0}, {362.5, 1099.0})->
-            quadratic({297.0, 1324.0}, {164.0, 1554.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 799;
-    }
-
-    static em asterisk(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({963.0, 1247.0});
-        pathSegment0.
-            line({604.0, 1053.0})->
-            line({963.0, 858.0})->
-            line({905.0, 760.0})->
-            line({569.0, 963.0})->
-            line({569.0, 586.0})->
-            line({455.0, 586.0})->
-            line({455.0, 963.0})->
-            line({119.0, 760.0})->
-            line({61.0, 858.0})->
-            line({420.0, 1053.0})->
-            line({61.0, 1247.0})->
-            line({119.0, 1346.0})->
-            line({455.0, 1143.0})->
-            line({455.0, 1520.0})->
-            line({569.0, 1520.0})->
-            line({569.0, 1143.0})->
-            line({905.0, 1346.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1024;
-    }
-
-    static em plus(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({942.0, 1284.0});
-        pathSegment0.
-            line({942.0, 727.0})->
-            line({1499.0, 727.0})->
-            line({1499.0, 557.0})->
-            line({942.0, 557.0})->
-            line({942.0, 0.0})->
-            line({774.0, 0.0})->
-            line({774.0, 557.0})->
-            line({217.0, 557.0})->
-            line({217.0, 727.0})->
-            line({774.0, 727.0})->
-            line({774.0, 1284.0})->
-            line({942.0, 1284.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1716;
-    }
-
-    static em comma(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({240.0, 254.0});
-        pathSegment0.
-            line({451.0, 254.0})->
-            line({451.0, 82.0})->
-            line({287.0, -238.0})->
-            line({158.0, -238.0})->
-            line({240.0, 82.0})->
-            line({240.0, 254.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 651;
-    }
-
-    static em hyphen(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({100.0, 643.0});
-        pathSegment0.
-            line({639.0, 643.0})->
-            line({639.0, 479.0})->
-            line({100.0, 479.0})->
-            line({100.0, 643.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 739;
-    }
-
-    static em period(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({219.0, 254.0});
-        pathSegment0.
-            line({430.0, 254.0})->
-            line({430.0, 0.0})->
-            line({219.0, 0.0})->
-            line({219.0, 254.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 651;
-    }
-
-    static em slash(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({520.0, 1493.0});
-        pathSegment0.
-            line({690.0, 1493.0})->
-            line({170.0, -190.0})->
-            line({0.0, -190.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 690;
-    }
-
-    static em zero(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({651.0, 1360.0});
-        pathSegment0.
-            quadratic({495.0, 1360.0}, {416.5, 1206.5})->
-            quadratic({338.0, 1053.0}, {338.0, 745.0})->
-            quadratic({338.0, 438.0}, {416.5, 284.5})->
-            quadratic({495.0, 131.0}, {651.0, 131.0})->
-            quadratic({808.0, 131.0}, {886.5, 284.5})->
-            quadratic({965.0, 438.0}, {965.0, 745.0})->
-            quadratic({965.0, 1053.0}, {886.5, 1206.5})->
-            quadratic({808.0, 1360.0}, {651.0, 1360.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({651.0, 1520.0});
-        pathSegment1.
-            quadratic({902.0, 1520.0}, {1034.5, 1321.5})->
-            quadratic({1167.0, 1123.0}, {1167.0, 745.0})->
-            quadratic({1167.0, 368.0}, {1034.5, 169.5})->
-            quadratic({902.0, -29.0}, {651.0, -29.0})->
-            quadratic({400.0, -29.0}, {267.5, 169.5})->
-            quadratic({135.0, 368.0}, {135.0, 745.0})->
-            quadratic({135.0, 1123.0}, {267.5, 1321.5})->
-            quadratic({400.0, 1520.0}, {651.0, 1520.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 2000;
-    }
-
-    static em one(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({254.0, 170.0});
-        pathSegment0.
-            line({584.0, 170.0})->
-            line({584.0, 1309.0})->
-            line({225.0, 1237.0})->
-            line({225.0, 1421.0})->
-            line({582.0, 1493.0})->
-            line({784.0, 1493.0})->
-            line({784.0, 170.0})->
-            line({1114.0, 170.0})->
-            line({1114.0, 0.0})->
-            line({254.0, 0.0})->
-            line({254.0, 170.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em two(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({393.0, 170.0});
-        pathSegment0.
-            line({1098.0, 170.0})->
-            line({1098.0, 0.0})->
-            line({150.0, 0.0})->
-            line({150.0, 170.0})->
-            quadratic({265.0, 289.0}, {463.5, 489.5})->
-            quadratic({662.0, 690.0}, {713.0, 748.0})->
-            quadratic({810.0, 857.0}, {848.5, 932.5})->
-            quadratic({887.0, 1008.0}, {887.0, 1081.0})->
-            quadratic({887.0, 1200.0}, {803.5, 1275.0})->
-            quadratic({720.0, 1350.0}, {586.0, 1350.0})->
-            quadratic({491.0, 1350.0}, {385.5, 1317.0})->
-            quadratic({280.0, 1284.0}, {160.0, 1217.0})->
-            line({160.0, 1421.0})->
-            quadratic({282.0, 1470.0}, {388.0, 1495.0})->
-            quadratic({494.0, 1520.0}, {582.0, 1520.0})->
-            quadratic({814.0, 1520.0}, {952.0, 1404.0})->
-            quadratic({1090.0, 1288.0}, {1090.0, 1094.0})->
-            quadratic({1090.0, 1002.0}, {1055.5, 919.5})->
-            quadratic({1021.0, 837.0}, {930.0, 725.0})->
-            quadratic({905.0, 696.0}, {771.0, 557.5})->
-            quadratic({637.0, 419.0}, {393.0, 170.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em three(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({831.0, 805.0});
-        pathSegment0.
-            quadratic({976.0, 774.0}, {1057.5, 676.0})->
-            quadratic({1139.0, 578.0}, {1139.0, 434.0})->
-            quadratic({1139.0, 213.0}, {987.0, 92.0})->
-            quadratic({835.0, -29.0}, {555.0, -29.0})->
-            quadratic({461.0, -29.0}, {361.5, -10.5})->
-            quadratic({262.0, 8.0}, {156.0, 45.0})->
-            line({156.0, 240.0})->
-            quadratic({240.0, 191.0}, {340.0, 166.0})->
-            quadratic({440.0, 141.0}, {549.0, 141.0})->
-            quadratic({739.0, 141.0}, {838.5, 216.0})->
-            quadratic({938.0, 291.0}, {938.0, 434.0})->
-            quadratic({938.0, 566.0}, {845.5, 640.5})->
-            quadratic({753.0, 715.0}, {588.0, 715.0})->
-            line({414.0, 715.0})->
-            line({414.0, 881.0})->
-            line({596.0, 881.0})->
-            quadratic({745.0, 881.0}, {824.0, 940.5})->
-            quadratic({903.0, 1000.0}, {903.0, 1112.0})->
-            quadratic({903.0, 1227.0}, {821.5, 1288.5})->
-            quadratic({740.0, 1350.0}, {588.0, 1350.0})->
-            quadratic({505.0, 1350.0}, {410.0, 1332.0})->
-            quadratic({315.0, 1314.0}, {201.0, 1276.0})->
-            line({201.0, 1456.0})->
-            quadratic({316.0, 1488.0}, {416.5, 1504.0})->
-            quadratic({517.0, 1520.0}, {606.0, 1520.0})->
-            quadratic({836.0, 1520.0}, {970.0, 1415.5})->
-            quadratic({1104.0, 1311.0}, {1104.0, 1133.0})->
-            quadratic({1104.0, 1009.0}, {1033.0, 923.5})->
-            quadratic({962.0, 838.0}, {831.0, 805.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em four(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({774.0, 1317.0});
-        pathSegment0.
-            line({264.0, 520.0})->
-            line({774.0, 520.0})->
-            line({774.0, 1317.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({721.0, 1493.0});
-        pathSegment1.
-            line({975.0, 1493.0})->
-            line({975.0, 520.0})->
-            line({1188.0, 520.0})->
-            line({1188.0, 352.0})->
-            line({975.0, 352.0})->
-            line({975.0, 0.0})->
-            line({774.0, 0.0})->
-            line({774.0, 352.0})->
-            line({100.0, 352.0})->
-            line({100.0, 547.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 2000;
-    }
-
-    static em five(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({221.0, 1493.0});
-        pathSegment0.
-            line({1014.0, 1493.0})->
-            line({1014.0, 1323.0})->
-            line({406.0, 1323.0})->
-            line({406.0, 957.0})->
-            quadratic({450.0, 972.0}, {494.0, 979.5})->
-            quadratic({538.0, 987.0}, {582.0, 987.0})->
-            quadratic({832.0, 987.0}, {978.0, 850.0})->
-            quadratic({1124.0, 713.0}, {1124.0, 479.0})->
-            quadratic({1124.0, 238.0}, {974.0, 104.5})->
-            quadratic({824.0, -29.0}, {551.0, -29.0})->
-            quadratic({457.0, -29.0}, {359.5, -13.0})->
-            quadratic({262.0, 3.0}, {158.0, 35.0})->
-            line({158.0, 238.0})->
-            quadratic({248.0, 189.0}, {344.0, 165.0})->
-            quadratic({440.0, 141.0}, {547.0, 141.0})->
-            quadratic({720.0, 141.0}, {821.0, 232.0})->
-            quadratic({922.0, 323.0}, {922.0, 479.0})->
-            quadratic({922.0, 635.0}, {821.0, 726.0})->
-            quadratic({720.0, 817.0}, {547.0, 817.0})->
-            quadratic({466.0, 817.0}, {385.5, 799.0})->
-            quadratic({305.0, 781.0}, {221.0, 743.0})->
-            line({221.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em six(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({676.0, 827.0});
-        pathSegment0.
-            quadratic({540.0, 827.0}, {460.5, 734.0})->
-            quadratic({381.0, 641.0}, {381.0, 479.0})->
-            quadratic({381.0, 318.0}, {460.5, 224.5})->
-            quadratic({540.0, 131.0}, {676.0, 131.0})->
-            quadratic({812.0, 131.0}, {891.5, 224.5})->
-            quadratic({971.0, 318.0}, {971.0, 479.0})->
-            quadratic({971.0, 641.0}, {891.5, 734.0})->
-            quadratic({812.0, 827.0}, {676.0, 827.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({1077.0, 1460.0});
-        pathSegment1.
-            line({1077.0, 1276.0})->
-            quadratic({1001.0, 1312.0}, {923.5, 1331.0})->
-            quadratic({846.0, 1350.0}, {770.0, 1350.0})->
-            quadratic({570.0, 1350.0}, {464.5, 1215.0})->
-            quadratic({359.0, 1080.0}, {344.0, 807.0})->
-            quadratic({403.0, 894.0}, {492.0, 940.5})->
-            quadratic({581.0, 987.0}, {688.0, 987.0})->
-            quadratic({913.0, 987.0}, {1043.5, 850.5})->
-            quadratic({1174.0, 714.0}, {1174.0, 479.0})->
-            quadratic({1174.0, 249.0}, {1038.0, 110.0})->
-            quadratic({902.0, -29.0}, {676.0, -29.0})->
-            quadratic({417.0, -29.0}, {280.0, 169.5})->
-            quadratic({143.0, 368.0}, {143.0, 745.0})->
-            quadratic({143.0, 1099.0}, {311.0, 1309.5})->
-            quadratic({479.0, 1520.0}, {762.0, 1520.0})->
-            quadratic({838.0, 1520.0}, {915.5, 1505.0})->
-            quadratic({993.0, 1490.0}, {1077.0, 1460.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 2000;
-    }
-
-    static em seven(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({168.0, 1493.0});
-        pathSegment0.
-            line({1128.0, 1493.0})->
-            line({1128.0, 1407.0})->
-            line({586.0, 0.0})->
-            line({375.0, 0.0})->
-            line({885.0, 1323.0})->
-            line({168.0, 1323.0})->
-            line({168.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em eight(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({651.0, 709.0});
-        pathSegment0.
-            quadratic({507.0, 709.0}, {424.5, 632.0})->
-            quadratic({342.0, 555.0}, {342.0, 420.0})->
-            quadratic({342.0, 285.0}, {424.5, 208.0})->
-            quadratic({507.0, 131.0}, {651.0, 131.0})->
-            quadratic({795.0, 131.0}, {878.0, 208.5})->
-            quadratic({961.0, 286.0}, {961.0, 420.0})->
-            quadratic({961.0, 555.0}, {878.5, 632.0})->
-            quadratic({796.0, 709.0}, {651.0, 709.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({449.0, 795.0});
-        pathSegment1.
-            quadratic({319.0, 827.0}, {246.5, 916.0})->
-            quadratic({174.0, 1005.0}, {174.0, 1133.0})->
-            quadratic({174.0, 1312.0}, {301.5, 1416.0})->
-            quadratic({429.0, 1520.0}, {651.0, 1520.0})->
-            quadratic({874.0, 1520.0}, {1001.0, 1416.0})->
-            quadratic({1128.0, 1312.0}, {1128.0, 1133.0})->
-            quadratic({1128.0, 1005.0}, {1055.5, 916.0})->
-            quadratic({983.0, 827.0}, {854.0, 795.0})->
-            quadratic({1000.0, 761.0}, {1081.5, 662.0})->
-            quadratic({1163.0, 563.0}, {1163.0, 420.0})->
-            quadratic({1163.0, 203.0}, {1030.5, 87.0})->
-            quadratic({898.0, -29.0}, {651.0, -29.0})->
-            quadratic({404.0, -29.0}, {271.5, 87.0})->
-            quadratic({139.0, 203.0}, {139.0, 420.0})->
-            quadratic({139.0, 563.0}, {221.0, 662.0})->
-            quadratic({303.0, 761.0}, {449.0, 795.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        Path pathSegment2 = Path::start({375.0, 1114.0});
-        pathSegment2.
-            quadratic({375.0, 998.0}, {447.5, 933.0})->
-            quadratic({520.0, 868.0}, {651.0, 868.0})->
-            quadratic({781.0, 868.0}, {854.5, 933.0})->
-            quadratic({928.0, 998.0}, {928.0, 1114.0})->
-            quadratic({928.0, 1230.0}, {854.5, 1295.0})->
-            quadratic({781.0, 1360.0}, {651.0, 1360.0})->
-            quadratic({520.0, 1360.0}, {447.5, 1295.0})->
-            quadratic({375.0, 1230.0}, {375.0, 1114.0})->
-            close(&pathSegment2);
-        pathSegment2.drawTo(box);
-
-        return 2000;
-    }
-
-    static em nine(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({225.0, 31.0});
-        pathSegment0.
-            line({225.0, 215.0})->
-            quadratic({301.0, 179.0}, {379.0, 160.0})->
-            quadratic({457.0, 141.0}, {532.0, 141.0})->
-            quadratic({732.0, 141.0}, {837.5, 275.5})->
-            quadratic({943.0, 410.0}, {958.0, 684.0})->
-            quadratic({900.0, 598.0}, {811.0, 552.0})->
-            quadratic({722.0, 506.0}, {614.0, 506.0})->
-            quadratic({390.0, 506.0}, {259.5, 641.5})->
-            quadratic({129.0, 777.0}, {129.0, 1012.0})->
-            quadratic({129.0, 1242.0}, {265.0, 1381.0})->
-            quadratic({401.0, 1520.0}, {627.0, 1520.0})->
-            quadratic({886.0, 1520.0}, {1022.5, 1321.5})->
-            quadratic({1159.0, 1123.0}, {1159.0, 745.0})->
-            quadratic({1159.0, 392.0}, {991.5, 181.5})->
-            quadratic({824.0, -29.0}, {541.0, -29.0})->
-            quadratic({465.0, -29.0}, {387.0, -14.0})->
-            quadratic({309.0, 1.0}, {225.0, 31.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({627.0, 664.0});
-        pathSegment1.
-            quadratic({763.0, 664.0}, {842.5, 757.0})->
-            quadratic({922.0, 850.0}, {922.0, 1012.0})->
-            quadratic({922.0, 1173.0}, {842.5, 1266.5})->
-            quadratic({763.0, 1360.0}, {627.0, 1360.0})->
-            quadratic({491.0, 1360.0}, {411.5, 1266.5})->
-            quadratic({332.0, 1173.0}, {332.0, 1012.0})->
-            quadratic({332.0, 850.0}, {411.5, 757.0})->
-            quadratic({491.0, 664.0}, {627.0, 664.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 2000;
-    }
-
-    static em colon(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({240.0, 254.0});
-        pathSegment0.
-            line({451.0, 254.0})->
-            line({451.0, 0.0})->
-            line({240.0, 0.0})->
-            line({240.0, 254.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({240.0, 1059.0});
-        pathSegment1.
-            line({451.0, 1059.0})->
-            line({451.0, 805.0})->
-            line({240.0, 805.0})->
-            line({240.0, 1059.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 690;
-    }
-
-    static em semicolon(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({240.0, 1059.0});
-        pathSegment0.
-            line({451.0, 1059.0})->
-            line({451.0, 805.0})->
-            line({240.0, 805.0})->
-            line({240.0, 1059.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({240.0, 254.0});
-        pathSegment1.
-            line({451.0, 254.0})->
-            line({451.0, 82.0})->
-            line({287.0, -238.0})->
-            line({158.0, -238.0})->
-            line({240.0, 82.0})->
-            line({240.0, 254.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 690;
-    }
-
-    static em less(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1499.0, 1008.0});
-        pathSegment0.
-            line({467.0, 641.0})->
-            line({1499.0, 276.0})->
-            line({1499.0, 94.0})->
-            line({217.0, 559.0})->
-            line({217.0, 725.0})->
-            line({1499.0, 1190.0})->
-            line({1499.0, 1008.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1716;
-    }
-
-    static em equal(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({217.0, 930.0});
-        pathSegment0.
-            line({1499.0, 930.0})->
-            line({1499.0, 762.0})->
-            line({217.0, 762.0})->
-            line({217.0, 930.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({217.0, 522.0});
-        pathSegment1.
-            line({1499.0, 522.0})->
-            line({1499.0, 352.0})->
-            line({217.0, 352.0})->
-            line({217.0, 522.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1716;
-    }
-
-    static em greater(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({217.0, 1008.0});
-        pathSegment0.
-            line({217.0, 1190.0})->
-            line({1499.0, 725.0})->
-            line({1499.0, 559.0})->
-            line({217.0, 94.0})->
-            line({217.0, 276.0})->
-            line({1247.0, 641.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1716;
-    }
-
-    static em question(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({391.0, 254.0});
-        pathSegment0.
-            line({594.0, 254.0})->
-            line({594.0, 0.0})->
-            line({391.0, 0.0})->
-            line({391.0, 254.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({588.0, 401.0});
-        pathSegment1.
-            line({397.0, 401.0})->
-            line({397.0, 555.0})->
-            quadratic({397.0, 656.0}, {425.0, 721.0})->
-            quadratic({453.0, 786.0}, {543.0, 872.0})->
-            line({633.0, 961.0})->
-            quadratic({690.0, 1014.0}, {715.5, 1061.0})->
-            quadratic({741.0, 1108.0}, {741.0, 1157.0})->
-            quadratic({741.0, 1246.0}, {675.5, 1301.0})->
-            quadratic({610.0, 1356.0}, {502.0, 1356.0})->
-            quadratic({423.0, 1356.0}, {333.5, 1321.0})->
-            quadratic({244.0, 1286.0}, {147.0, 1219.0})->
-            line({147.0, 1407.0})->
-            quadratic({241.0, 1464.0}, {337.5, 1492.0})->
-            quadratic({434.0, 1520.0}, {537.0, 1520.0})->
-            quadratic({721.0, 1520.0}, {832.5, 1423.0})->
-            quadratic({944.0, 1326.0}, {944.0, 1167.0})->
-            quadratic({944.0, 1091.0}, {908.0, 1022.5})->
-            quadratic({872.0, 954.0}, {782.0, 868.0})->
-            line({694.0, 782.0})->
-            quadratic({647.0, 735.0}, {627.5, 708.5})->
-            quadratic({608.0, 682.0}, {600.0, 657.0})->
-            quadratic({594.0, 636.0}, {591.0, 606.0})->
-            quadratic({588.0, 576.0}, {588.0, 524.0})->
-            line({588.0, 401.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1087;
-    }
-
-    static em at(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({762.0, 537.0});
-        pathSegment0.
-            quadratic({762.0, 394.0}, {833.0, 312.5})->
-            quadratic({904.0, 231.0}, {1028.0, 231.0})->
-            quadratic({1151.0, 231.0}, {1221.5, 313.0})->
-            quadratic({1292.0, 395.0}, {1292.0, 537.0})->
-            quadratic({1292.0, 677.0}, {1220.0, 759.5})->
-            quadratic({1148.0, 842.0}, {1026.0, 842.0})->
-            quadratic({905.0, 842.0}, {833.5, 760.0})->
-            quadratic({762.0, 678.0}, {762.0, 537.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({1307.0, 238.0});
-        pathSegment1.
-            quadratic({1247.0, 161.0}, {1169.5, 124.5})->
-            quadratic({1092.0, 88.0}, {989.0, 88.0})->
-            quadratic({817.0, 88.0}, {709.5, 212.5})->
-            quadratic({602.0, 337.0}, {602.0, 537.0})->
-            quadratic({602.0, 737.0}, {710.0, 862.0})->
-            quadratic({818.0, 987.0}, {989.0, 987.0})->
-            quadratic({1092.0, 987.0}, {1170.0, 949.5})->
-            quadratic({1248.0, 912.0}, {1307.0, 836.0})->
-            line({1307.0, 967.0})->
-            line({1450.0, 967.0})->
-            line({1450.0, 231.0})->
-            quadratic({1596.0, 253.0}, {1678.5, 364.5})->
-            quadratic({1761.0, 476.0}, {1761.0, 653.0})->
-            quadratic({1761.0, 760.0}, {1729.5, 854.0})->
-            quadratic({1698.0, 948.0}, {1634.0, 1028.0})->
-            quadratic({1530.0, 1159.0}, {1380.5, 1228.5})->
-            quadratic({1231.0, 1298.0}, {1055.0, 1298.0})->
-            quadratic({932.0, 1298.0}, {819.0, 1265.5})->
-            quadratic({706.0, 1233.0}, {610.0, 1169.0})->
-            quadratic({453.0, 1067.0}, {364.5, 901.5})->
-            quadratic({276.0, 736.0}, {276.0, 543.0})->
-            quadratic({276.0, 384.0}, {333.5, 245.0})->
-            quadratic({391.0, 106.0}, {500.0, 0.0})->
-            quadratic({605.0, -104.0}, {743.0, -158.5})->
-            quadratic({881.0, -213.0}, {1038.0, -213.0})->
-            quadratic({1167.0, -213.0}, {1291.5, -169.5})->
-            quadratic({1416.0, -126.0}, {1520.0, -45.0})->
-            line({1610.0, -156.0})->
-            quadratic({1485.0, -253.0}, {1337.5, -304.5})->
-            quadratic({1190.0, -356.0}, {1038.0, -356.0})->
-            quadratic({853.0, -356.0}, {689.0, -290.5})->
-            quadratic({525.0, -225.0}, {397.0, -100.0})->
-            quadratic({269.0, 25.0}, {202.0, 189.5})->
-            quadratic({135.0, 354.0}, {135.0, 543.0})->
-            quadratic({135.0, 725.0}, {203.0, 890.0})->
-            quadratic({271.0, 1055.0}, {397.0, 1180.0})->
-            quadratic({526.0, 1307.0}, {695.0, 1374.5})->
-            quadratic({864.0, 1442.0}, {1053.0, 1442.0})->
-            quadratic({1265.0, 1442.0}, {1446.5, 1355.0})->
-            quadratic({1628.0, 1268.0}, {1751.0, 1108.0})->
-            quadratic({1826.0, 1010.0}, {1865.5, 895.0})->
-            quadratic({1905.0, 780.0}, {1905.0, 657.0})->
-            quadratic({1905.0, 394.0}, {1746.0, 242.0})->
-            quadratic({1587.0, 90.0}, {1307.0, 84.0})->
-            line({1307.0, 238.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 2048;
-    }
-
-    static em A(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({700.0, 1294.0});
-        pathSegment0.
-            line({426.0, 551.0})->
-            line({975.0, 551.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({586.0, 1493.0});
-        pathSegment1.
-            line({815.0, 1493.0})->
-            line({1384.0, 0.0})->
-            line({1174.0, 0.0})->
-            line({1038.0, 383.0})->
-            line({365.0, 383.0})->
-            line({229.0, 0.0})->
-            line({16.0, 0.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1401;
-    }
-
-    static em B(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({403.0, 713.0});
-        pathSegment0.
-            line({403.0, 166.0})->
-            line({727.0, 166.0})->
-            quadratic({890.0, 166.0}, {968.5, 233.5})->
-            quadratic({1047.0, 301.0}, {1047.0, 440.0})->
-            quadratic({1047.0, 580.0}, {968.5, 646.5})->
-            quadratic({890.0, 713.0}, {727.0, 713.0})->
-            line({403.0, 713.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({403.0, 1327.0});
-        pathSegment1.
-            line({403.0, 877.0})->
-            line({702.0, 877.0})->
-            quadratic({850.0, 877.0}, {922.5, 932.5})->
-            quadratic({995.0, 988.0}, {995.0, 1102.0})->
-            quadratic({995.0, 1215.0}, {922.5, 1271.0})->
-            quadratic({850.0, 1327.0}, {702.0, 1327.0})->
-            line({403.0, 1327.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        Path pathSegment2 = Path::start({201.0, 1493.0});
-        pathSegment2.
-            line({717.0, 1493.0})->
-            quadratic({948.0, 1493.0}, {1073.0, 1397.0})->
-            quadratic({1198.0, 1301.0}, {1198.0, 1124.0})->
-            quadratic({1198.0, 987.0}, {1134.0, 906.0})->
-            quadratic({1070.0, 825.0}, {946.0, 805.0})->
-            quadratic({1095.0, 773.0}, {1177.5, 671.5})->
-            quadratic({1260.0, 570.0}, {1260.0, 418.0})->
-            quadratic({1260.0, 218.0}, {1124.0, 109.0})->
-            quadratic({988.0, 0.0}, {737.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment2);
-        pathSegment2.drawTo(box);
-
-        return 1405;
-    }
-
-    static em C(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1319.0, 1378.0});
-        pathSegment0.
-            line({1319.0, 1165.0})->
-            quadratic({1217.0, 1260.0}, {1101.5, 1307.0})->
-            quadratic({986.0, 1354.0}, {856.0, 1354.0})->
-            quadratic({600.0, 1354.0}, {464.0, 1197.5})->
-            quadratic({328.0, 1041.0}, {328.0, 745.0})->
-            quadratic({328.0, 450.0}, {464.0, 293.5})->
-            quadratic({600.0, 137.0}, {856.0, 137.0})->
-            quadratic({986.0, 137.0}, {1101.5, 184.0})->
-            quadratic({1217.0, 231.0}, {1319.0, 326.0})->
-            line({1319.0, 115.0})->
-            quadratic({1213.0, 43.0}, {1094.5, 7.0})->
-            quadratic({976.0, -29.0}, {844.0, -29.0})->
-            quadratic({505.0, -29.0}, {310.0, 178.5})->
-            quadratic({115.0, 386.0}, {115.0, 745.0})->
-            quadratic({115.0, 1105.0}, {310.0, 1312.5})->
-            quadratic({505.0, 1520.0}, {844.0, 1520.0})->
-            quadratic({978.0, 1520.0}, {1096.5, 1484.5})->
-            quadratic({1215.0, 1449.0}, {1319.0, 1378.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1430;
-    }
-
-    static em D(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({403.0, 1327.0});
-        pathSegment0.
-            line({403.0, 166.0})->
-            line({647.0, 166.0})->
-            quadratic({956.0, 166.0}, {1099.5, 306.0})->
-            quadratic({1243.0, 446.0}, {1243.0, 748.0})->
-            quadratic({1243.0, 1048.0}, {1099.5, 1187.5})->
-            quadratic({956.0, 1327.0}, {647.0, 1327.0})->
-            line({403.0, 1327.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({201.0, 1493.0});
-        pathSegment1.
-            line({616.0, 1493.0})->
-            quadratic({1050.0, 1493.0}, {1253.0, 1312.5})->
-            quadratic({1456.0, 1132.0}, {1456.0, 748.0})->
-            quadratic({1456.0, 362.0}, {1252.0, 181.0})->
-            quadratic({1048.0, 0.0}, {616.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1577;
-    }
-
-    static em E(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({1145.0, 1493.0})->
-            line({1145.0, 1323.0})->
-            line({403.0, 1323.0})->
-            line({403.0, 881.0})->
-            line({1114.0, 881.0})->
-            line({1114.0, 711.0})->
-            line({403.0, 711.0})->
-            line({403.0, 170.0})->
-            line({1163.0, 170.0})->
-            line({1163.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1294;
-    }
-
-    static em F(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({1059.0, 1493.0})->
-            line({1059.0, 1323.0})->
-            line({403.0, 1323.0})->
-            line({403.0, 883.0})->
-            line({995.0, 883.0})->
-            line({995.0, 713.0})->
-            line({403.0, 713.0})->
-            line({403.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1178;
-    }
-
-    static em G(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1219.0, 213.0});
-        pathSegment0.
-            line({1219.0, 614.0})->
-            line({889.0, 614.0})->
-            line({889.0, 780.0})->
-            line({1419.0, 780.0})->
-            line({1419.0, 139.0})->
-            quadratic({1302.0, 56.0}, {1161.0, 13.5})->
-            quadratic({1020.0, -29.0}, {860.0, -29.0})->
-            quadratic({510.0, -29.0}, {312.5, 175.5})->
-            quadratic({115.0, 380.0}, {115.0, 745.0})->
-            quadratic({115.0, 1111.0}, {312.5, 1315.5})->
-            quadratic({510.0, 1520.0}, {860.0, 1520.0})->
-            quadratic({1006.0, 1520.0}, {1137.5, 1484.0})->
-            quadratic({1269.0, 1448.0}, {1380.0, 1378.0})->
-            line({1380.0, 1163.0})->
-            quadratic({1268.0, 1258.0}, {1142.0, 1306.0})->
-            quadratic({1016.0, 1354.0}, {877.0, 1354.0})->
-            quadratic({603.0, 1354.0}, {465.5, 1201.0})->
-            quadratic({328.0, 1048.0}, {328.0, 745.0})->
-            quadratic({328.0, 443.0}, {465.5, 290.0})->
-            quadratic({603.0, 137.0}, {877.0, 137.0})->
-            quadratic({984.0, 137.0}, {1068.0, 155.5})->
-            quadratic({1152.0, 174.0}, {1219.0, 213.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1587;
-    }
-
-    static em H(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({403.0, 1493.0})->
-            line({403.0, 881.0})->
-            line({1137.0, 881.0})->
-            line({1137.0, 1493.0})->
-            line({1339.0, 1493.0})->
-            line({1339.0, 0.0})->
-            line({1137.0, 0.0})->
-            line({1137.0, 711.0})->
-            line({403.0, 711.0})->
-            line({403.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1540;
-    }
-
-    static em I(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({403.0, 1493.0})->
-            line({403.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 604;
-    }
-
-    static em J(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({403.0, 1493.0})->
-            line({403.0, 104.0})->
-            quadratic({403.0, -166.0}, {300.5, -288.0})->
-            quadratic({198.0, -410.0}, {-29.0, -410.0})->
-            line({-106.0, -410.0})->
-            line({-106.0, -240.0})->
-            line({-43.0, -240.0})->
-            quadratic({91.0, -240.0}, {146.0, -165.0})->
-            quadratic({201.0, -90.0}, {201.0, 104.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 604;
-    }
-
-    static em K(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({403.0, 1493.0})->
-            line({403.0, 862.0})->
-            line({1073.0, 1493.0})->
-            line({1333.0, 1493.0})->
-            line({592.0, 797.0})->
-            line({1386.0, 0.0})->
-            line({1120.0, 0.0})->
-            line({403.0, 719.0})->
-            line({403.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1343;
-    }
-
-    static em L(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({403.0, 1493.0})->
-            line({403.0, 170.0})->
-            line({1130.0, 170.0})->
-            line({1130.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1141;
-    }
-
-    static em M(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({502.0, 1493.0})->
-            line({883.0, 477.0})->
-            line({1266.0, 1493.0})->
-            line({1567.0, 1493.0})->
-            line({1567.0, 0.0})->
-            line({1370.0, 0.0})->
-            line({1370.0, 1311.0})->
-            line({985.0, 287.0})->
-            line({782.0, 287.0})->
-            line({397.0, 1311.0})->
-            line({397.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1767;
-    }
-
-    static em N(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({201.0, 1493.0});
-        pathSegment0.
-            line({473.0, 1493.0})->
-            line({1135.0, 244.0})->
-            line({1135.0, 1493.0})->
-            line({1331.0, 1493.0})->
-            line({1331.0, 0.0})->
-            line({1059.0, 0.0})->
-            line({397.0, 1249.0})->
-            line({397.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1532;
-    }
-
-    static em O(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({807.0, 1356.0});
-        pathSegment0.
-            quadratic({587.0, 1356.0}, {457.5, 1192.0})->
-            quadratic({328.0, 1028.0}, {328.0, 745.0})->
-            quadratic({328.0, 463.0}, {457.5, 299.0})->
-            quadratic({587.0, 135.0}, {807.0, 135.0})->
-            quadratic({1027.0, 135.0}, {1155.5, 299.0})->
-            quadratic({1284.0, 463.0}, {1284.0, 745.0})->
-            quadratic({1284.0, 1028.0}, {1155.5, 1192.0})->
-            quadratic({1027.0, 1356.0}, {807.0, 1356.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({807.0, 1520.0});
-        pathSegment1.
-            quadratic({1121.0, 1520.0}, {1309.0, 1309.5})->
-            quadratic({1497.0, 1099.0}, {1497.0, 745.0})->
-            quadratic({1497.0, 392.0}, {1309.0, 181.5})->
-            quadratic({1121.0, -29.0}, {807.0, -29.0})->
-            quadratic({492.0, -29.0}, {303.5, 181.0})->
-            quadratic({115.0, 391.0}, {115.0, 745.0})->
-            quadratic({115.0, 1099.0}, {303.5, 1309.5})->
-            quadratic({492.0, 1520.0}, {807.0, 1520.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1612;
-    }
-
-    static em P(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({403.0, 1327.0});
-        pathSegment0.
-            line({403.0, 766.0})->
-            line({657.0, 766.0})->
-            quadratic({798.0, 766.0}, {875.0, 839.0})->
-            quadratic({952.0, 912.0}, {952.0, 1047.0})->
-            quadratic({952.0, 1181.0}, {875.0, 1254.0})->
-            quadratic({798.0, 1327.0}, {657.0, 1327.0})->
-            line({403.0, 1327.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({201.0, 1493.0});
-        pathSegment1.
-            line({657.0, 1493.0})->
-            quadratic({908.0, 1493.0}, {1036.5, 1379.5})->
-            quadratic({1165.0, 1266.0}, {1165.0, 1047.0})->
-            quadratic({1165.0, 826.0}, {1036.5, 713.0})->
-            quadratic({908.0, 600.0}, {657.0, 600.0})->
-            line({403.0, 600.0})->
-            line({403.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1235;
-    }
-
-    static em Q(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({807.0, 1356.0});
-        pathSegment0.
-            quadratic({587.0, 1356.0}, {457.5, 1192.0})->
-            quadratic({328.0, 1028.0}, {328.0, 745.0})->
-            quadratic({328.0, 463.0}, {457.5, 299.0})->
-            quadratic({587.0, 135.0}, {807.0, 135.0})->
-            quadratic({1027.0, 135.0}, {1155.5, 299.0})->
-            quadratic({1284.0, 463.0}, {1284.0, 745.0})->
-            quadratic({1284.0, 1028.0}, {1155.5, 1192.0})->
-            quadratic({1027.0, 1356.0}, {807.0, 1356.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({1090.0, 27.0});
-        pathSegment1.
-            line({1356.0, -264.0})->
-            line({1112.0, -264.0})->
-            line({891.0, -25.0})->
-            quadratic({858.0, -27.0}, {840.5, -28.0})->
-            quadratic({823.0, -29.0}, {807.0, -29.0})->
-            quadratic({492.0, -29.0}, {303.5, 181.5})->
-            quadratic({115.0, 392.0}, {115.0, 745.0})->
-            quadratic({115.0, 1099.0}, {303.5, 1309.5})->
-            quadratic({492.0, 1520.0}, {807.0, 1520.0})->
-            quadratic({1121.0, 1520.0}, {1309.0, 1309.5})->
-            quadratic({1497.0, 1099.0}, {1497.0, 745.0})->
-            quadratic({1497.0, 485.0}, {1392.5, 300.0})->
-            quadratic({1288.0, 115.0}, {1090.0, 27.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1612;
-    }
-
-    static em R(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({909.0, 700.0});
-        pathSegment0.
-            quadratic({974.0, 678.0}, {1035.5, 606.0})->
-            quadratic({1097.0, 534.0}, {1159.0, 408.0})->
-            line({1364.0, 0.0})->
-            line({1147.0, 0.0})->
-            line({956.0, 383.0})->
-            quadratic({882.0, 533.0}, {812.5, 582.0})->
-            quadratic({743.0, 631.0}, {623.0, 631.0})->
-            line({403.0, 631.0})->
-            line({403.0, 0.0})->
-            line({201.0, 0.0})->
-            line({201.0, 1493.0})->
-            line({657.0, 1493.0})->
-            quadratic({913.0, 1493.0}, {1039.0, 1386.0})->
-            quadratic({1165.0, 1279.0}, {1165.0, 1063.0})->
-            quadratic({1165.0, 922.0}, {1099.5, 829.0})->
-            quadratic({1034.0, 736.0}, {909.0, 700.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({403.0, 1327.0});
-        pathSegment1.
-            line({403.0, 797.0})->
-            line({657.0, 797.0})->
-            quadratic({803.0, 797.0}, {877.5, 864.5})->
-            quadratic({952.0, 932.0}, {952.0, 1063.0})->
-            quadratic({952.0, 1194.0}, {877.5, 1260.5})->
-            quadratic({803.0, 1327.0}, {657.0, 1327.0})->
-            line({403.0, 1327.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1423;
-    }
-
-    static em S(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1096.0, 1444.0});
-        pathSegment0.
-            line({1096.0, 1247.0})->
-            quadratic({981.0, 1302.0}, {879.0, 1329.0})->
-            quadratic({777.0, 1356.0}, {682.0, 1356.0})->
-            quadratic({517.0, 1356.0}, {427.5, 1292.0})->
-            quadratic({338.0, 1228.0}, {338.0, 1110.0})->
-            quadratic({338.0, 1011.0}, {397.5, 960.5})->
-            quadratic({457.0, 910.0}, {623.0, 879.0})->
-            line({745.0, 854.0})->
-            quadratic({971.0, 811.0}, {1078.5, 702.5})->
-            quadratic({1186.0, 594.0}, {1186.0, 412.0})->
-            quadratic({1186.0, 195.0}, {1040.5, 83.0})->
-            quadratic({895.0, -29.0}, {614.0, -29.0})->
-            quadratic({508.0, -29.0}, {388.5, -5.0})->
-            quadratic({269.0, 19.0}, {141.0, 66.0})->
-            line({141.0, 274.0})->
-            quadratic({264.0, 205.0}, {382.0, 170.0})->
-            quadratic({500.0, 135.0}, {614.0, 135.0})->
-            quadratic({787.0, 135.0}, {881.0, 203.0})->
-            quadratic({975.0, 271.0}, {975.0, 397.0})->
-            quadratic({975.0, 507.0}, {907.5, 569.0})->
-            quadratic({840.0, 631.0}, {686.0, 662.0})->
-            line({563.0, 686.0})->
-            quadratic({337.0, 731.0}, {236.0, 827.0})->
-            quadratic({135.0, 923.0}, {135.0, 1094.0})->
-            quadratic({135.0, 1292.0}, {274.5, 1406.0})->
-            quadratic({414.0, 1520.0}, {659.0, 1520.0})->
-            quadratic({764.0, 1520.0}, {873.0, 1501.0})->
-            quadratic({982.0, 1482.0}, {1096.0, 1444.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1300;
-    }
-
-    static em T(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({-6.0, 1493.0});
-        pathSegment0.
-            line({1257.0, 1493.0})->
-            line({1257.0, 1323.0})->
-            line({727.0, 1323.0})->
-            line({727.0, 0.0})->
-            line({524.0, 0.0})->
-            line({524.0, 1323.0})->
-            line({-6.0, 1323.0})->
-            line({-6.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1251;
-    }
-
-    static em U(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({178.0, 1493.0});
-        pathSegment0.
-            line({381.0, 1493.0})->
-            line({381.0, 586.0})->
-            quadratic({381.0, 346.0}, {468.0, 240.5})->
-            quadratic({555.0, 135.0}, {750.0, 135.0})->
-            quadratic({944.0, 135.0}, {1031.0, 240.5})->
-            quadratic({1118.0, 346.0}, {1118.0, 586.0})->
-            line({1118.0, 1493.0})->
-            line({1321.0, 1493.0})->
-            line({1321.0, 561.0})->
-            quadratic({1321.0, 269.0}, {1176.5, 120.0})->
-            quadratic({1032.0, -29.0}, {750.0, -29.0})->
-            quadratic({467.0, -29.0}, {322.5, 120.0})->
-            quadratic({178.0, 269.0}, {178.0, 561.0})->
-            line({178.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1499;
-    }
-
-    static em V(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({586.0, 0.0});
-        pathSegment0.
-            line({16.0, 1493.0})->
-            line({227.0, 1493.0})->
-            line({700.0, 236.0})->
-            line({1174.0, 1493.0})->
-            line({1384.0, 1493.0})->
-            line({815.0, 0.0})->
-            line({586.0, 0.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1401;
-    }
-
-    static em W(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({68.0, 1493.0});
-        pathSegment0.
-            line({272.0, 1493.0})->
-            line({586.0, 231.0})->
-            line({899.0, 1493.0})->
-            line({1126.0, 1493.0})->
-            line({1440.0, 231.0})->
-            line({1753.0, 1493.0})->
-            line({1958.0, 1493.0})->
-            line({1583.0, 0.0})->
-            line({1329.0, 0.0})->
-            line({1014.0, 1296.0})->
-            line({696.0, 0.0})->
-            line({442.0, 0.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2025;
-    }
-
-    static em X(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({129.0, 1493.0});
-        pathSegment0.
-            line({346.0, 1493.0})->
-            line({717.0, 938.0})->
-            line({1090.0, 1493.0})->
-            line({1307.0, 1493.0})->
-            line({827.0, 776.0})->
-            line({1339.0, 0.0})->
-            line({1122.0, 0.0})->
-            line({702.0, 635.0})->
-            line({279.0, 0.0})->
-            line({61.0, 0.0})->
-            line({594.0, 797.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1403;
-    }
-
-    static em Y(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({-4.0, 1493.0});
-        pathSegment0.
-            line({213.0, 1493.0})->
-            line({627.0, 879.0})->
-            line({1038.0, 1493.0})->
-            line({1255.0, 1493.0})->
-            line({727.0, 711.0})->
-            line({727.0, 0.0})->
-            line({524.0, 0.0})->
-            line({524.0, 711.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1251;
-    }
-
-    static em Z(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({115.0, 1493.0});
-        pathSegment0.
-            line({1288.0, 1493.0})->
-            line({1288.0, 1339.0})->
-            line({344.0, 170.0})->
-            line({1311.0, 170.0})->
-            line({1311.0, 0.0})->
-            line({92.0, 0.0})->
-            line({92.0, 154.0})->
-            line({1036.0, 1323.0})->
-            line({115.0, 1323.0})->
-            line({115.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1403;
-    }
-
-    static em bracketleft(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({176.0, 1556.0});
-        pathSegment0.
-            line({600.0, 1556.0})->
-            line({600.0, 1413.0})->
-            line({360.0, 1413.0})->
-            line({360.0, -127.0})->
-            line({600.0, -127.0})->
-            line({600.0, -270.0})->
-            line({176.0, -270.0})->
-            line({176.0, 1556.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 799;
-    }
-
-    static em backslash(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({170.0, 1493.0});
-        pathSegment0.
-            line({690.0, -190.0})->
-            line({520.0, -190.0})->
-            line({0.0, 1493.0})->
-            line({170.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 690;
-    }
-
-    static em bracketright(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({623.0, 1556.0});
-        pathSegment0.
-            line({623.0, -270.0})->
-            line({199.0, -270.0})->
-            line({199.0, -127.0})->
-            line({438.0, -127.0})->
-            line({438.0, 1413.0})->
-            line({199.0, 1413.0})->
-            line({199.0, 1556.0})->
-            line({623.0, 1556.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 799;
-    }
-
-    static em asciicircum(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({956.0, 1493.0});
-        pathSegment0.
-            line({1499.0, 936.0})->
-            line({1298.0, 936.0})->
-            line({858.0, 1331.0})->
-            line({418.0, 936.0})->
-            line({217.0, 936.0})->
-            line({760.0, 1493.0})->
-            line({956.0, 1493.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1716;
-    }
-
-    static em underscore(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1044.0, -340.0});
-        pathSegment0.
-            line({1044.0, -483.0})->
-            line({-20.0, -483.0})->
-            line({-20.0, -340.0})->
-            line({1044.0, -340.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1024;
-    }
-
-    static em grave(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({367.0, 1638.0});
-        pathSegment0.
-            line({649.0, 1264.0})->
-            line({496.0, 1264.0})->
-            line({170.0, 1638.0})->
-            line({367.0, 1638.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1024;
-    }
-
-    static em a(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({702.0, 563.0});
-        pathSegment0.
-            quadratic({479.0, 563.0}, {393.0, 512.0})->
-            quadratic({307.0, 461.0}, {307.0, 338.0})->
-            quadratic({307.0, 240.0}, {371.5, 182.5})->
-            quadratic({436.0, 125.0}, {547.0, 125.0})->
-            quadratic({700.0, 125.0}, {792.5, 233.5})->
-            quadratic({885.0, 342.0}, {885.0, 522.0})->
-            line({885.0, 563.0})->
-            line({702.0, 563.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({1069.0, 639.0});
-        pathSegment1.
-            line({1069.0, 0.0})->
-            line({885.0, 0.0})->
-            line({885.0, 170.0})->
-            quadratic({822.0, 68.0}, {728.0, 19.5})->
-            quadratic({634.0, -29.0}, {498.0, -29.0})->
-            quadratic({326.0, -29.0}, {224.5, 67.5})->
-            quadratic({123.0, 164.0}, {123.0, 326.0})->
-            quadratic({123.0, 515.0}, {249.5, 611.0})->
-            quadratic({376.0, 707.0}, {627.0, 707.0})->
-            line({885.0, 707.0})->
-            line({885.0, 725.0})->
-            quadratic({885.0, 852.0}, {801.5, 921.5})->
-            quadratic({718.0, 991.0}, {567.0, 991.0})->
-            quadratic({471.0, 991.0}, {380.0, 968.0})->
-            quadratic({289.0, 945.0}, {205.0, 899.0})->
-            line({205.0, 1069.0})->
-            quadratic({306.0, 1108.0}, {401.0, 1127.5})->
-            quadratic({496.0, 1147.0}, {586.0, 1147.0})->
-            quadratic({829.0, 1147.0}, {949.0, 1021.0})->
-            quadratic({1069.0, 895.0}, {1069.0, 639.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1255;
-    }
-
-    static em b(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({997.0, 559.0});
-        pathSegment0.
-            quadratic({997.0, 762.0}, {913.5, 877.5})->
-            quadratic({830.0, 993.0}, {684.0, 993.0})->
-            quadratic({538.0, 993.0}, {454.5, 877.5})->
-            quadratic({371.0, 762.0}, {371.0, 559.0})->
-            quadratic({371.0, 356.0}, {454.5, 240.5})->
-            quadratic({538.0, 125.0}, {684.0, 125.0})->
-            quadratic({830.0, 125.0}, {913.5, 240.5})->
-            quadratic({997.0, 356.0}, {997.0, 559.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({371.0, 950.0});
-        pathSegment1.
-            quadratic({429.0, 1050.0}, {517.5, 1098.5})->
-            quadratic({606.0, 1147.0}, {729.0, 1147.0})->
-            quadratic({933.0, 1147.0}, {1060.5, 985.0})->
-            quadratic({1188.0, 823.0}, {1188.0, 559.0})->
-            quadratic({1188.0, 295.0}, {1060.5, 133.0})->
-            quadratic({933.0, -29.0}, {729.0, -29.0})->
-            quadratic({606.0, -29.0}, {517.5, 19.5})->
-            quadratic({429.0, 68.0}, {371.0, 168.0})->
-            line({371.0, 0.0})->
-            line({186.0, 0.0})->
-            line({186.0, 1556.0})->
-            line({371.0, 1556.0})->
-            line({371.0, 950.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1300;
-    }
-
-    static em c(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({999.0, 1077.0});
-        pathSegment0.
-            line({999.0, 905.0})->
-            quadratic({921.0, 948.0}, {842.5, 969.5})->
-            quadratic({764.0, 991.0}, {684.0, 991.0})->
-            quadratic({505.0, 991.0}, {406.0, 877.5})->
-            quadratic({307.0, 764.0}, {307.0, 559.0})->
-            quadratic({307.0, 354.0}, {406.0, 240.5})->
-            quadratic({505.0, 127.0}, {684.0, 127.0})->
-            quadratic({764.0, 127.0}, {842.5, 148.5})->
-            quadratic({921.0, 170.0}, {999.0, 213.0})->
-            line({999.0, 43.0})->
-            quadratic({922.0, 7.0}, {839.5, -11.0})->
-            quadratic({757.0, -29.0}, {664.0, -29.0})->
-            quadratic({411.0, -29.0}, {262.0, 130.0})->
-            quadratic({113.0, 289.0}, {113.0, 559.0})->
-            quadratic({113.0, 833.0}, {263.5, 990.0})->
-            quadratic({414.0, 1147.0}, {676.0, 1147.0})->
-            quadratic({761.0, 1147.0}, {842.0, 1129.5})->
-            quadratic({923.0, 1112.0}, {999.0, 1077.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1126;
-    }
-
-    static em d(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({930.0, 950.0});
-        pathSegment0.
-            line({930.0, 1556.0})->
-            line({1114.0, 1556.0})->
-            line({1114.0, 0.0})->
-            line({930.0, 0.0})->
-            line({930.0, 168.0})->
-            quadratic({872.0, 68.0}, {783.5, 19.5})->
-            quadratic({695.0, -29.0}, {571.0, -29.0})->
-            quadratic({368.0, -29.0}, {240.5, 133.0})->
-            quadratic({113.0, 295.0}, {113.0, 559.0})->
-            quadratic({113.0, 823.0}, {240.5, 985.0})->
-            quadratic({368.0, 1147.0}, {571.0, 1147.0})->
-            quadratic({695.0, 1147.0}, {783.5, 1098.5})->
-            quadratic({872.0, 1050.0}, {930.0, 950.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({303.0, 559.0});
-        pathSegment1.
-            quadratic({303.0, 356.0}, {386.5, 240.5})->
-            quadratic({470.0, 125.0}, {616.0, 125.0})->
-            quadratic({762.0, 125.0}, {846.0, 240.5})->
-            quadratic({930.0, 356.0}, {930.0, 559.0})->
-            quadratic({930.0, 762.0}, {846.0, 877.5})->
-            quadratic({762.0, 993.0}, {616.0, 993.0})->
-            quadratic({470.0, 993.0}, {386.5, 877.5})->
-            quadratic({303.0, 762.0}, {303.0, 559.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1300;
-    }
-
-    static em e(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1151.0, 606.0});
-        pathSegment0.
-            line({1151.0, 516.0})->
-            line({305.0, 516.0})->
-            quadratic({317.0, 326.0}, {419.5, 226.5})->
-            quadratic({522.0, 127.0}, {705.0, 127.0})->
-            quadratic({811.0, 127.0}, {910.5, 153.0})->
-            quadratic({1010.0, 179.0}, {1108.0, 231.0})->
-            line({1108.0, 57.0})->
-            quadratic({1009.0, 15.0}, {905.0, -7.0})->
-            quadratic({801.0, -29.0}, {694.0, -29.0})->
-            quadratic({426.0, -29.0}, {269.5, 127.0})->
-            quadratic({113.0, 283.0}, {113.0, 549.0})->
-            quadratic({113.0, 824.0}, {261.5, 985.5})->
-            quadratic({410.0, 1147.0}, {662.0, 1147.0})->
-            quadratic({888.0, 1147.0}, {1019.5, 1001.5})->
-            quadratic({1151.0, 856.0}, {1151.0, 606.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({967.0, 660.0});
-        pathSegment1.
-            quadratic({965.0, 811.0}, {882.5, 901.0})->
-            quadratic({800.0, 991.0}, {664.0, 991.0})->
-            quadratic({510.0, 991.0}, {417.5, 904.0})->
-            quadratic({325.0, 817.0}, {311.0, 659.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1260;
-    }
-
-    static em f(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({760.0, 1556.0});
-        pathSegment0.
-            line({760.0, 1403.0})->
-            line({584.0, 1403.0})->
-            quadratic({485.0, 1403.0}, {446.5, 1363.0})->
-            quadratic({408.0, 1323.0}, {408.0, 1219.0})->
-            line({408.0, 1120.0})->
-            line({711.0, 1120.0})->
-            line({711.0, 977.0})->
-            line({408.0, 977.0})->
-            line({408.0, 0.0})->
-            line({223.0, 0.0})->
-            line({223.0, 977.0})->
-            line({47.0, 977.0})->
-            line({47.0, 1120.0})->
-            line({223.0, 1120.0})->
-            line({223.0, 1198.0})->
-            quadratic({223.0, 1385.0}, {310.0, 1470.5})->
-            quadratic({397.0, 1556.0}, {586.0, 1556.0})->
-            line({760.0, 1556.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 721;
-    }
-
-    static em g(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({930.0, 573.0});
-        pathSegment0.
-            quadratic({930.0, 773.0}, {847.5, 883.0})->
-            quadratic({765.0, 993.0}, {616.0, 993.0})->
-            quadratic({468.0, 993.0}, {385.5, 883.0})->
-            quadratic({303.0, 773.0}, {303.0, 573.0})->
-            quadratic({303.0, 374.0}, {385.5, 264.0})->
-            quadratic({468.0, 154.0}, {616.0, 154.0})->
-            quadratic({765.0, 154.0}, {847.5, 264.0})->
-            quadratic({930.0, 374.0}, {930.0, 573.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({1114.0, 139.0});
-        pathSegment1.
-            quadratic({1114.0, -147.0}, {987.0, -286.5})->
-            quadratic({860.0, -426.0}, {598.0, -426.0})->
-            quadratic({501.0, -426.0}, {415.0, -411.5})->
-            quadratic({329.0, -397.0}, {248.0, -367.0})->
-            line({248.0, -188.0})->
-            quadratic({329.0, -232.0}, {408.0, -253.0})->
-            quadratic({487.0, -274.0}, {569.0, -274.0})->
-            quadratic({750.0, -274.0}, {840.0, -179.5})->
-            quadratic({930.0, -85.0}, {930.0, 106.0})->
-            line({930.0, 197.0})->
-            quadratic({873.0, 98.0}, {784.0, 49.0})->
-            quadratic({695.0, 0.0}, {571.0, 0.0})->
-            quadratic({365.0, 0.0}, {239.0, 157.0})->
-            quadratic({113.0, 314.0}, {113.0, 573.0})->
-            quadratic({113.0, 833.0}, {239.0, 990.0})->
-            quadratic({365.0, 1147.0}, {571.0, 1147.0})->
-            quadratic({695.0, 1147.0}, {784.0, 1098.0})->
-            quadratic({873.0, 1049.0}, {930.0, 950.0})->
-            line({930.0, 1120.0})->
-            line({1114.0, 1120.0})->
-            line({1114.0, 139.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1300;
-    }
-
-    static em h(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1124.0, 676.0});
-        pathSegment0.
-            line({1124.0, 0.0})->
-            line({940.0, 0.0})->
-            line({940.0, 670.0})->
-            quadratic({940.0, 829.0}, {878.0, 908.0})->
-            quadratic({816.0, 987.0}, {692.0, 987.0})->
-            quadratic({543.0, 987.0}, {457.0, 892.0})->
-            quadratic({371.0, 797.0}, {371.0, 633.0})->
-            line({371.0, 0.0})->
-            line({186.0, 0.0})->
-            line({186.0, 1556.0})->
-            line({371.0, 1556.0})->
-            line({371.0, 946.0})->
-            quadratic({437.0, 1047.0}, {526.5, 1097.0})->
-            quadratic({616.0, 1147.0}, {733.0, 1147.0})->
-            quadratic({926.0, 1147.0}, {1025.0, 1027.5})->
-            quadratic({1124.0, 908.0}, {1124.0, 676.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1298;
-    }
-
-    static em i(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({193.0, 1120.0});
-        pathSegment0.
-            line({377.0, 1120.0})->
-            line({377.0, 0.0})->
-            line({193.0, 0.0})->
-            line({193.0, 1120.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({193.0, 1556.0});
-        pathSegment1.
-            line({377.0, 1556.0})->
-            line({377.0, 1323.0})->
-            line({193.0, 1323.0})->
-            line({193.0, 1556.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 569;
-    }
-
-    static em j(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({193.0, 1120.0});
-        pathSegment0.
-            line({377.0, 1120.0})->
-            line({377.0, -20.0})->
-            quadratic({377.0, -234.0}, {295.5, -330.0})->
-            quadratic({214.0, -426.0}, {33.0, -426.0})->
-            line({-37.0, -426.0})->
-            line({-37.0, -270.0})->
-            line({12.0, -270.0})->
-            quadratic({117.0, -270.0}, {155.0, -221.5})->
-            quadratic({193.0, -173.0}, {193.0, -20.0})->
-            line({193.0, 1120.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({193.0, 1556.0});
-        pathSegment1.
-            line({377.0, 1556.0})->
-            line({377.0, 1323.0})->
-            line({193.0, 1323.0})->
-            line({193.0, 1556.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 569;
-    }
-
-    static em k(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({186.0, 1556.0});
-        pathSegment0.
-            line({371.0, 1556.0})->
-            line({371.0, 637.0})->
-            line({920.0, 1120.0})->
-            line({1155.0, 1120.0})->
-            line({561.0, 596.0})->
-            line({1180.0, 0.0})->
-            line({940.0, 0.0})->
-            line({371.0, 547.0})->
-            line({371.0, 0.0})->
-            line({186.0, 0.0})->
-            line({186.0, 1556.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1186;
-    }
-
-    static em l(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({193.0, 1556.0});
-        pathSegment0.
-            line({377.0, 1556.0})->
-            line({377.0, 0.0})->
-            line({193.0, 0.0})->
-            line({193.0, 1556.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 569;
-    }
-
-    static em m(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1065.0, 905.0});
-        pathSegment0.
-            quadratic({1134.0, 1029.0}, {1230.0, 1088.0})->
-            quadratic({1326.0, 1147.0}, {1456.0, 1147.0})->
-            quadratic({1631.0, 1147.0}, {1726.0, 1024.5})->
-            quadratic({1821.0, 902.0}, {1821.0, 676.0})->
-            line({1821.0, 0.0})->
-            line({1636.0, 0.0})->
-            line({1636.0, 670.0})->
-            quadratic({1636.0, 831.0}, {1579.0, 909.0})->
-            quadratic({1522.0, 987.0}, {1405.0, 987.0})->
-            quadratic({1262.0, 987.0}, {1179.0, 892.0})->
-            quadratic({1096.0, 797.0}, {1096.0, 633.0})->
-            line({1096.0, 0.0})->
-            line({911.0, 0.0})->
-            line({911.0, 670.0})->
-            quadratic({911.0, 832.0}, {854.0, 909.5})->
-            quadratic({797.0, 987.0}, {678.0, 987.0})->
-            quadratic({537.0, 987.0}, {454.0, 891.5})->
-            quadratic({371.0, 796.0}, {371.0, 633.0})->
-            line({371.0, 0.0})->
-            line({186.0, 0.0})->
-            line({186.0, 1120.0})->
-            line({371.0, 1120.0})->
-            line({371.0, 946.0})->
-            quadratic({434.0, 1049.0}, {522.0, 1098.0})->
-            quadratic({610.0, 1147.0}, {731.0, 1147.0})->
-            quadratic({853.0, 1147.0}, {938.5, 1085.0})->
-            quadratic({1024.0, 1023.0}, {1065.0, 905.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1995;
-    }
-
-    static em n(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1124.0, 676.0});
-        pathSegment0.
-            line({1124.0, 0.0})->
-            line({940.0, 0.0})->
-            line({940.0, 670.0})->
-            quadratic({940.0, 829.0}, {878.0, 908.0})->
-            quadratic({816.0, 987.0}, {692.0, 987.0})->
-            quadratic({543.0, 987.0}, {457.0, 892.0})->
-            quadratic({371.0, 797.0}, {371.0, 633.0})->
-            line({371.0, 0.0})->
-            line({186.0, 0.0})->
-            line({186.0, 1120.0})->
-            line({371.0, 1120.0})->
-            line({371.0, 946.0})->
-            quadratic({437.0, 1047.0}, {526.5, 1097.0})->
-            quadratic({616.0, 1147.0}, {733.0, 1147.0})->
-            quadratic({926.0, 1147.0}, {1025.0, 1027.5})->
-            quadratic({1124.0, 908.0}, {1124.0, 676.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1298;
-    }
-
-    static em o(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({627.0, 991.0});
-        pathSegment0.
-            quadratic({479.0, 991.0}, {393.0, 875.5})->
-            quadratic({307.0, 760.0}, {307.0, 559.0})->
-            quadratic({307.0, 358.0}, {392.5, 242.5})->
-            quadratic({478.0, 127.0}, {627.0, 127.0})->
-            quadratic({774.0, 127.0}, {860.0, 243.0})->
-            quadratic({946.0, 359.0}, {946.0, 559.0})->
-            quadratic({946.0, 758.0}, {860.0, 874.5})->
-            quadratic({774.0, 991.0}, {627.0, 991.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({627.0, 1147.0});
-        pathSegment1.
-            quadratic({867.0, 1147.0}, {1004.0, 991.0})->
-            quadratic({1141.0, 835.0}, {1141.0, 559.0})->
-            quadratic({1141.0, 284.0}, {1004.0, 127.5})->
-            quadratic({867.0, -29.0}, {627.0, -29.0})->
-            quadratic({386.0, -29.0}, {249.5, 127.5})->
-            quadratic({113.0, 284.0}, {113.0, 559.0})->
-            quadratic({113.0, 835.0}, {249.5, 991.0})->
-            quadratic({386.0, 1147.0}, {627.0, 1147.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1253;
-    }
-
-    static em p(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({371.0, 168.0});
-        pathSegment0.
-            line({371.0, -426.0})->
-            line({186.0, -426.0})->
-            line({186.0, 1120.0})->
-            line({371.0, 1120.0})->
-            line({371.0, 950.0})->
-            quadratic({429.0, 1050.0}, {517.5, 1098.5})->
-            quadratic({606.0, 1147.0}, {729.0, 1147.0})->
-            quadratic({933.0, 1147.0}, {1060.5, 985.0})->
-            quadratic({1188.0, 823.0}, {1188.0, 559.0})->
-            quadratic({1188.0, 295.0}, {1060.5, 133.0})->
-            quadratic({933.0, -29.0}, {729.0, -29.0})->
-            quadratic({606.0, -29.0}, {517.5, 19.5})->
-            quadratic({429.0, 68.0}, {371.0, 168.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({997.0, 559.0});
-        pathSegment1.
-            quadratic({997.0, 762.0}, {913.5, 877.5})->
-            quadratic({830.0, 993.0}, {684.0, 993.0})->
-            quadratic({538.0, 993.0}, {454.5, 877.5})->
-            quadratic({371.0, 762.0}, {371.0, 559.0})->
-            quadratic({371.0, 356.0}, {454.5, 240.5})->
-            quadratic({538.0, 125.0}, {684.0, 125.0})->
-            quadratic({830.0, 125.0}, {913.5, 240.5})->
-            quadratic({997.0, 356.0}, {997.0, 559.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1300;
-    }
-
-    static em q(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({303.0, 559.0});
-        pathSegment0.
-            quadratic({303.0, 356.0}, {386.5, 240.5})->
-            quadratic({470.0, 125.0}, {616.0, 125.0})->
-            quadratic({762.0, 125.0}, {846.0, 240.5})->
-            quadratic({930.0, 356.0}, {930.0, 559.0})->
-            quadratic({930.0, 762.0}, {846.0, 877.5})->
-            quadratic({762.0, 993.0}, {616.0, 993.0})->
-            quadratic({470.0, 993.0}, {386.5, 877.5})->
-            quadratic({303.0, 762.0}, {303.0, 559.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({930.0, 168.0});
-        pathSegment1.
-            quadratic({872.0, 68.0}, {783.5, 19.5})->
-            quadratic({695.0, -29.0}, {571.0, -29.0})->
-            quadratic({368.0, -29.0}, {240.5, 133.0})->
-            quadratic({113.0, 295.0}, {113.0, 559.0})->
-            quadratic({113.0, 823.0}, {240.5, 985.0})->
-            quadratic({368.0, 1147.0}, {571.0, 1147.0})->
-            quadratic({695.0, 1147.0}, {783.5, 1098.5})->
-            quadratic({872.0, 1050.0}, {930.0, 950.0})->
-            line({930.0, 1120.0})->
-            line({1114.0, 1120.0})->
-            line({1114.0, -426.0})->
-            line({930.0, -426.0})->
-            line({930.0, 168.0})->
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1300;
-    }
-
-    static em r(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({842.0, 948.0});
-        pathSegment0.
-            quadratic({811.0, 966.0}, {774.5, 974.5})->
-            quadratic({738.0, 983.0}, {694.0, 983.0})->
-            quadratic({538.0, 983.0}, {454.5, 881.5})->
-            quadratic({371.0, 780.0}, {371.0, 590.0})->
-            line({371.0, 0.0})->
-            line({186.0, 0.0})->
-            line({186.0, 1120.0})->
-            line({371.0, 1120.0})->
-            line({371.0, 946.0})->
-            quadratic({429.0, 1048.0}, {522.0, 1097.5})->
-            quadratic({615.0, 1147.0}, {748.0, 1147.0})->
-            quadratic({767.0, 1147.0}, {790.0, 1144.5})->
-            quadratic({813.0, 1142.0}, {841.0, 1137.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 842;
-    }
-
-    static em s(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({907.0, 1087.0});
-        pathSegment0.
-            line({907.0, 913.0})->
-            quadratic({829.0, 953.0}, {745.0, 973.0})->
-            quadratic({661.0, 993.0}, {571.0, 993.0})->
-            quadratic({434.0, 993.0}, {365.5, 951.0})->
-            quadratic({297.0, 909.0}, {297.0, 825.0})->
-            quadratic({297.0, 761.0}, {346.0, 724.5})->
-            quadratic({395.0, 688.0}, {543.0, 655.0})->
-            line({606.0, 641.0})->
-            quadratic({802.0, 599.0}, {884.5, 522.5})->
-            quadratic({967.0, 446.0}, {967.0, 309.0})->
-            quadratic({967.0, 153.0}, {843.5, 62.0})->
-            quadratic({720.0, -29.0}, {504.0, -29.0})->
-            quadratic({414.0, -29.0}, {316.5, -11.5})->
-            quadratic({219.0, 6.0}, {111.0, 41.0})->
-            line({111.0, 231.0})->
-            quadratic({213.0, 178.0}, {312.0, 151.5})->
-            quadratic({411.0, 125.0}, {508.0, 125.0})->
-            quadratic({638.0, 125.0}, {708.0, 169.5})->
-            quadratic({778.0, 214.0}, {778.0, 295.0})->
-            quadratic({778.0, 370.0}, {727.5, 410.0})->
-            quadratic({677.0, 450.0}, {506.0, 487.0})->
-            line({442.0, 502.0})->
-            quadratic({271.0, 538.0}, {195.0, 612.5})->
-            quadratic({119.0, 687.0}, {119.0, 817.0})->
-            quadratic({119.0, 975.0}, {231.0, 1061.0})->
-            quadratic({343.0, 1147.0}, {549.0, 1147.0})->
-            quadratic({651.0, 1147.0}, {741.0, 1132.0})->
-            quadratic({831.0, 1117.0}, {907.0, 1087.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1067;
-    }
-
-    static em t(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({375.0, 1438.0});
-        pathSegment0.
-            line({375.0, 1120.0})->
-            line({754.0, 1120.0})->
-            line({754.0, 977.0})->
-            line({375.0, 977.0})->
-            line({375.0, 369.0})->
-            quadratic({375.0, 232.0}, {412.5, 193.0})->
-            quadratic({450.0, 154.0}, {565.0, 154.0})->
-            line({754.0, 154.0})->
-            line({754.0, 0.0})->
-            line({565.0, 0.0})->
-            quadratic({352.0, 0.0}, {271.0, 79.5})->
-            quadratic({190.0, 159.0}, {190.0, 369.0})->
-            line({190.0, 977.0})->
-            line({55.0, 977.0})->
-            line({55.0, 1120.0})->
-            line({190.0, 1120.0})->
-            line({190.0, 1438.0})->
-            line({375.0, 1438.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 803;
-    }
-
-    static em u(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({174.0, 442.0});
-        pathSegment0.
-            line({174.0, 1120.0})->
-            line({358.0, 1120.0})->
-            line({358.0, 449.0})->
-            quadratic({358.0, 290.0}, {420.0, 210.5})->
-            quadratic({482.0, 131.0}, {606.0, 131.0})->
-            quadratic({755.0, 131.0}, {841.5, 226.0})->
-            quadratic({928.0, 321.0}, {928.0, 485.0})->
-            line({928.0, 1120.0})->
-            line({1112.0, 1120.0})->
-            line({1112.0, 0.0})->
-            line({928.0, 0.0})->
-            line({928.0, 172.0})->
-            quadratic({861.0, 70.0}, {772.5, 20.5})->
-            quadratic({684.0, -29.0}, {567.0, -29.0})->
-            quadratic({374.0, -29.0}, {274.0, 91.0})->
-            quadratic({174.0, 211.0}, {174.0, 442.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        Path pathSegment1 = Path::start({637.0, 1147.0});
-        pathSegment1.
-            close(&pathSegment1);
-        pathSegment1.drawTo(box);
-
-        return 1298;
-    }
-
-    static em v(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({61.0, 1120.0});
-        pathSegment0.
-            line({256.0, 1120.0})->
-            line({606.0, 180.0})->
-            line({956.0, 1120.0})->
-            line({1151.0, 1120.0})->
-            line({731.0, 0.0})->
-            line({481.0, 0.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1212;
-    }
-
-    static em w(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({86.0, 1120.0});
-        pathSegment0.
-            line({270.0, 1120.0})->
-            line({500.0, 246.0})->
-            line({729.0, 1120.0})->
-            line({946.0, 1120.0})->
-            line({1176.0, 246.0})->
-            line({1405.0, 1120.0})->
-            line({1589.0, 1120.0})->
-            line({1296.0, 0.0})->
-            line({1079.0, 0.0})->
-            line({838.0, 918.0})->
-            line({596.0, 0.0})->
-            line({379.0, 0.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1675;
-    }
-
-    static em x(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1124.0, 1120.0});
-        pathSegment0.
-            line({719.0, 575.0})->
-            line({1145.0, 0.0})->
-            line({928.0, 0.0})->
-            line({602.0, 440.0})->
-            line({276.0, 0.0})->
-            line({59.0, 0.0})->
-            line({494.0, 586.0})->
-            line({96.0, 1120.0})->
-            line({313.0, 1120.0})->
-            line({610.0, 721.0})->
-            line({907.0, 1120.0})->
-            line({1124.0, 1120.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1212;
-    }
-
-    static em y(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({659.0, -104.0});
-        pathSegment0.
-            quadratic({581.0, -304.0}, {507.0, -365.0})->
-            quadratic({433.0, -426.0}, {309.0, -426.0})->
-            line({162.0, -426.0})->
-            line({162.0, -272.0})->
-            line({270.0, -272.0})->
-            quadratic({346.0, -272.0}, {388.0, -236.0})->
-            quadratic({430.0, -200.0}, {481.0, -66.0})->
-            line({514.0, 18.0})->
-            line({61.0, 1120.0})->
-            line({256.0, 1120.0})->
-            line({606.0, 244.0})->
-            line({956.0, 1120.0})->
-            line({1151.0, 1120.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1212;
-    }
-
-    static em z(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({113.0, 1120.0});
-        pathSegment0.
-            line({987.0, 1120.0})->
-            line({987.0, 952.0})->
-            line({295.0, 147.0})->
-            line({987.0, 147.0})->
-            line({987.0, 0.0})->
-            line({88.0, 0.0})->
-            line({88.0, 168.0})->
-            line({780.0, 973.0})->
-            line({113.0, 973.0})->
-            line({113.0, 1120.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1075;
-    }
-
-    static em braceleft(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1047.0, -190.0});
-        pathSegment0.
-            line({1047.0, -334.0})->
-            line({985.0, -334.0})->
-            quadratic({736.0, -334.0}, {651.5, -260.0})->
-            quadratic({567.0, -186.0}, {567.0, 35.0})->
-            line({567.0, 274.0})->
-            quadratic({567.0, 425.0}, {513.0, 483.0})->
-            quadratic({459.0, 541.0}, {317.0, 541.0})->
-            line({256.0, 541.0})->
-            line({256.0, 684.0})->
-            line({317.0, 684.0})->
-            quadratic({460.0, 684.0}, {513.5, 741.5})->
-            quadratic({567.0, 799.0}, {567.0, 948.0})->
-            line({567.0, 1188.0})->
-            quadratic({567.0, 1409.0}, {651.5, 1482.5})->
-            quadratic({736.0, 1556.0}, {985.0, 1556.0})->
-            line({1047.0, 1556.0})->
-            line({1047.0, 1413.0})->
-            line({979.0, 1413.0})->
-            quadratic({838.0, 1413.0}, {795.0, 1369.0})->
-            quadratic({752.0, 1325.0}, {752.0, 1184.0})->
-            line({752.0, 936.0})->
-            quadratic({752.0, 779.0}, {706.5, 708.0})->
-            quadratic({661.0, 637.0}, {551.0, 612.0})->
-            quadratic({662.0, 585.0}, {707.0, 514.0})->
-            quadratic({752.0, 443.0}, {752.0, 287.0})->
-            line({752.0, 39.0})->
-            quadratic({752.0, -102.0}, {795.0, -146.0})->
-            quadratic({838.0, -190.0}, {979.0, -190.0})->
-            line({1047.0, -190.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em bar(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({430.0, 1565.0});
-        pathSegment0.
-            line({430.0, -483.0})->
-            line({260.0, -483.0})->
-            line({260.0, 1565.0})->
-            line({430.0, 1565.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 690;
-    }
-
-    static em braceright(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({256.0, -190.0});
-        pathSegment0.
-            line({326.0, -190.0})->
-            quadratic({466.0, -190.0}, {508.5, -147.0})->
-            quadratic({551.0, -104.0}, {551.0, 39.0})->
-            line({551.0, 287.0})->
-            quadratic({551.0, 443.0}, {596.0, 514.0})->
-            quadratic({641.0, 585.0}, {752.0, 612.0})->
-            quadratic({641.0, 637.0}, {596.0, 708.0})->
-            quadratic({551.0, 779.0}, {551.0, 936.0})->
-            line({551.0, 1184.0})->
-            quadratic({551.0, 1326.0}, {508.5, 1369.5})->
-            quadratic({466.0, 1413.0}, {326.0, 1413.0})->
-            line({256.0, 1413.0})->
-            line({256.0, 1556.0})->
-            line({319.0, 1556.0})->
-            quadratic({568.0, 1556.0}, {651.5, 1482.5})->
-            quadratic({735.0, 1409.0}, {735.0, 1188.0})->
-            line({735.0, 948.0})->
-            quadratic({735.0, 799.0}, {789.0, 741.5})->
-            quadratic({843.0, 684.0}, {985.0, 684.0})->
-            line({1047.0, 684.0})->
-            line({1047.0, 541.0})->
-            line({985.0, 541.0})->
-            quadratic({843.0, 541.0}, {789.0, 483.0})->
-            quadratic({735.0, 425.0}, {735.0, 274.0})->
-            line({735.0, 35.0})->
-            quadratic({735.0, -186.0}, {651.5, -260.0})->
-            quadratic({568.0, -334.0}, {319.0, -334.0})->
-            line({256.0, -334.0})->
-            line({256.0, -190.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 2000;
-    }
-
-    static em asciitilde(EmBox* box)
-    {
-        Path pathSegment0 = Path::start({1499.0, 817.0});
-        pathSegment0.
-            line({1499.0, 639.0})->
-            quadratic({1394.0, 560.0}, {1304.5, 526.0})->
-            quadratic({1215.0, 492.0}, {1118.0, 492.0})->
-            quadratic({1008.0, 492.0}, {862.0, 551.0})->
-            quadratic({851.0, 555.0}, {846.0, 557.0})->
-            quadratic({839.0, 560.0}, {824.0, 565.0})->
-            quadratic({669.0, 627.0}, {575.0, 627.0})->
-            quadratic({487.0, 627.0}, {401.0, 588.5})->
-            quadratic({315.0, 550.0}, {217.0, 467.0})->
-            line({217.0, 645.0})->
-            quadratic({322.0, 724.0}, {411.5, 758.5})->
-            quadratic({501.0, 793.0}, {598.0, 793.0})->
-            quadratic({708.0, 793.0}, {855.0, 733.0})->
-            line({870.0, 727.0})->
-            quadratic({878.0, 724.0}, {892.0, 719.0})->
-            quadratic({1047.0, 657.0}, {1141.0, 657.0})->
-            quadratic({1227.0, 657.0}, {1311.5, 695.0})->
-            quadratic({1396.0, 733.0}, {1499.0, 817.0})->
-            close(&pathSegment0);
-        pathSegment0.drawTo(box);
-
-        return 1716;
-    }
-
-    static em nonbreakingspace(EmBox* box)
-    {
-        return 651;
-    }
-
-};
 
 int main() {
     gpio_init(LED_PIN);
